Move timer to core
