+++ Sprocket/Core/timer.cpp
@@ -1,4 +1,4 @@
-#include "Stopwatch.h"
+#include "timer.h"
 
 namespace spkt {
